Two bug fixes.
get_connection() would automatically connect to an unencrypted network.  Even if it was not saved in wifi.dat.  It would ignore whatever network credentials were saved in wifi.dat.

handle_configure() would not properly decode a password entered by the user via the configuration page.  Therefore, the password saved for connecting to the user's desired wifi network would be saved with eroneous pluses and URL character codes.  This was fixed by using the micropython version of urllib.unquote_plus().